+++ acceptance/config/asset_manager.go
@@ -12,7 +12,7 @@ import (
 	"testing"
 
 	acceptanceOS "github.com/buildpacks/pack/acceptance/os"
-	"github.com/buildpacks/pack/internal/blob"
+	"github.com/buildpacks/pack/blob"
 	"github.com/buildpacks/pack/internal/builder"
 	"github.com/buildpacks/pack/internal/config"
 	"github.com/buildpacks/pack/internal/style"
