+++ blob/testdata/lifecycle/analyzer
