+++ blob/testdata/lifecycle/detector
