+++ blob/testdata/lifecycle/exporter
