+++ blob/testdata/lifecycle/launcher
