+++ image/fetcher_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/sclevine/spec/report"
 
 	pubcfg "github.com/buildpacks/pack/config"
-	"github.com/buildpacks/pack/internal/image"
+	"github.com/buildpacks/pack/image"
 	"github.com/buildpacks/pack/internal/logging"
 	h "github.com/buildpacks/pack/testhelpers"
 )
