+++ inspect_builder.go
@@ -5,9 +5,9 @@ import (
 
 	pubbldr "github.com/buildpacks/pack/builder"
 
+	"github.com/buildpacks/pack/image"
 	"github.com/buildpacks/pack/internal/builder"
 	"github.com/buildpacks/pack/internal/dist"
-	"github.com/buildpacks/pack/internal/image"
 )
 
 // BuilderInfo is a collection of metadata describing a builder created using pack.
