+++ inspect_buildpack_test.go
@@ -11,14 +11,14 @@ import (
 
 	"github.com/pkg/errors"
 
+	"github.com/buildpacks/pack/image"
 	cfg "github.com/buildpacks/pack/internal/config"
-	"github.com/buildpacks/pack/internal/image"
 
 	"github.com/google/go-containerregistry/pkg/v1/empty"
 	"github.com/google/go-containerregistry/pkg/v1/layout"
 	"github.com/google/go-containerregistry/pkg/v1/mutate"
 
-	"github.com/buildpacks/pack/internal/blob"
+	"github.com/buildpacks/pack/blob"
 	"github.com/buildpacks/pack/internal/buildpack"
 	"github.com/buildpacks/pack/pkg/archive"
