+++ inspect_image.go
@@ -11,8 +11,8 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/buildpacks/pack/config"
+	"github.com/buildpacks/pack/image"
 	"github.com/buildpacks/pack/internal/dist"
-	"github.com/buildpacks/pack/internal/image"
 )
 
 // ImageInfo is a collection of metadata describing
