+++ internal/builder/lifecycle_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/sclevine/spec"
 	"github.com/sclevine/spec/report"
 
-	"github.com/buildpacks/pack/internal/blob"
+	"github.com/buildpacks/pack/blob"
 	"github.com/buildpacks/pack/internal/builder"
 	h "github.com/buildpacks/pack/testhelpers"
 )
