+++ internal/buildpackage/oci_layout_package.go
@@ -12,7 +12,7 @@ import (
 	v1 "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/pkg/errors"
 
-	blob2 "github.com/buildpacks/pack/internal/blob"
+	blob2 "github.com/buildpacks/pack/blob"
 	"github.com/buildpacks/pack/internal/dist"
 	"github.com/buildpacks/pack/internal/style"
 	"github.com/buildpacks/pack/pkg/archive"
