+++ internal/buildpackage/oci_layout_package_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/buildpacks/lifecycle/api"
 
-	"github.com/buildpacks/pack/internal/blob"
+	"github.com/buildpacks/pack/blob"
 	"github.com/buildpacks/pack/internal/buildpackage"
 	"github.com/buildpacks/pack/internal/dist"
 	"github.com/buildpacks/pack/internal/fakes"
