+++ internal/fakes/fake_image_fetcher.go
@@ -8,7 +8,7 @@ import (
 	"github.com/buildpacks/imgutil"
 	"github.com/pkg/errors"
 
-	"github.com/buildpacks/pack/internal/image"
+	"github.com/buildpacks/pack/image"
 )
 
 type FetchArgs struct {
