+++ package_buildpack_test.go
@@ -19,14 +19,14 @@ import (
 	"github.com/sclevine/spec/report"
 
 	"github.com/buildpacks/pack"
+	"github.com/buildpacks/pack/blob"
 	pubbldpkg "github.com/buildpacks/pack/buildpackage"
 	pubcfg "github.com/buildpacks/pack/config"
-	"github.com/buildpacks/pack/internal/blob"
+	"github.com/buildpacks/pack/image"
 	"github.com/buildpacks/pack/internal/buildpackage"
 	cfg "github.com/buildpacks/pack/internal/config"
 	"github.com/buildpacks/pack/internal/dist"
 	ifakes "github.com/buildpacks/pack/internal/fakes"
-	"github.com/buildpacks/pack/internal/image"
 	"github.com/buildpacks/pack/internal/logging"
 	"github.com/buildpacks/pack/internal/paths"
 	h "github.com/buildpacks/pack/testhelpers"
